feat: extend extranonce1 and extranonce2 to 8 bytes each
Expand EXTRANONCE1_SIZE and EXTRANONCE2_SIZE from 4 to 8 bytes.
Total extranonce space grows from 8 to 16 bytes, giving miners
significantly more nonce search space before requiring a new job.

fix(db): propagate extranonce parse errors instead of unwrap

In the Result-returning fetch path, replace .unwrap() with proper error propagation using TupleAttributeParsingError for extranonce1
and extranonce2 hex parsing from DB.

fix(test): randomize extra_nonce_1 in create_test_bead

Replace hardcoded 42 with rand::random::<u64>() to match extra_nonce_2
and be consistent with emit_bead() in test_utils.rs.

+++ node/src/utils/test_utils.rs
@@ -127,8 +127,8 @@ pub mod test_utility_functions {
     }
 
     pub struct TestUnCommittedMetadataBuilder {
-        extra_nonce_1: u32,
-        extra_nonce_2: u32,
+        extra_nonce_1: u64,
+        extra_nonce_2: u64,
         broadcast_timestamp: Option<Time>,
         signature: Option<Signature>,
     }
@@ -144,7 +144,7 @@ pub mod test_utility_functions {
             }
         }
 
-        pub fn extra_nonce(mut self, nonce_1: u32, nonce_2: u32) -> Self {
+        pub fn extra_nonce(mut self, nonce_1: u64, nonce_2: u64) -> Self {
             self.extra_nonce_1 = nonce_1;
             self.extra_nonce_2 = nonce_2;
             self
@@ -348,8 +348,8 @@ pub mod test_utility_functions {
             .transactions(vec![])
             .build();
 
-        let extra_nonce_1 = rand::random::<u32>();
-        let extra_nonce_2 = rand::random::<u32>();
+        let extra_nonce_1 = rand::random::<u64>();
+        let extra_nonce_2 = rand::random::<u64>();
 
         let secp = Secp256k1::new();
